feat: сохраняются события (добавление в друзья, удаление из друзей, добавление и удаление лайка, добавление удаление и обновление отзыва)

+++ src/main/java/ru/yandex/practicum/filmorate/controller/DirectorController.java
@@ -36,7 +36,7 @@ class DirectorController {
     }
 
     @DeleteMapping("/{id}")
-    public void removeDirectorById(@PathVariable int id) {
+    public void removeDirectorById(@PathVariable long id) {
         directorService.removeDirector(id);
     }
 }
